# schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Eventglass API
  version: 0.1.0
  description: Eventglass setup, session, project, Issue, log search, record detail, and system contracts.
servers:
  - url: /
paths:
...
  /api/explore/aggregate:
    post:
      operationId: aggregate
      summary: Aggregate an authorized read snapshot
      security:
      - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregateRequest'
      responses:
        '200':
          description: Exact native aggregation for the shared read snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '410':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
        '504':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: eventglass_session
...
  responses:
    Error:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    Int64String:
      type: string
      pattern: "^-?[0-9]+$"
      description: Signed 64-bit integer encoded as a decimal string.
    PositiveDecimalString:
      type: string
      pattern: "^[1-9][0-9]*$"
      description: Positive 64-bit integer encoded as a decimal string.
    NonNegativeDecimalString:
      type: string
      pattern: "^[0-9]+$"
      description: Non-negative 64-bit integer encoded as a decimal string.
...
    SearchProjectId:
      oneOf:
        - $ref: "#/components/schemas/PositiveDecimalString"
        - type: integer
          format: int64
          minimum: 1
      description: A positive project ID accepted as a decimal string or JSON integer.
    SearchFilters:
      type: object
      additionalProperties: false
      properties:
        kinds:
          type: array
          maxItems: 64
          items:
            type: string
            enum: [log, error]
        services:
          $ref: "#/components/schemas/SearchFilterValues"
        levels:
          $ref: "#/components/schemas/SearchFilterValues"
        environments:
          $ref: "#/components/schemas/SearchFilterValues"
        releases:
          $ref: "#/components/schemas/SearchFilterValues"
        loggers:
          $ref: "#/components/schemas/SearchFilterValues"
    SearchFilterValues:
      type: array
      maxItems: 64
      items:
        type: string
        maxLength: 8192
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            retryable:
              type: boolean
    AggregateMetricInput:
      oneOf:
      - type: object
        additionalProperties: false
        required:
        - op
        properties:
          op:
            type: string
            enum:
            - count
          name: &id001
            type: string
            minLength: 1
            maxLength: 128
      - type: object
        additionalProperties: false
        required:
        - op
        - field
        properties:
          op:
            type: string
            enum:
            - sum
            - min
            - max
            - avg
          name: *id001
          field:
            type: string
            pattern: ^attributes\..+
            maxLength: 523
    AggregateRequest:
      type: object
      additionalProperties: false
      required:
      - start
      - end
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/SearchProjectId'
          maxItems: 1000
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        query:
          type: string
          maxLength: 8192
        filters:
          $ref: '#/components/schemas/SearchFilters'
        read_token:
          type: string
          maxLength: 12288
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/AggregateMetricInput'
          maxItems: 8
        group_by:
          type: array
          items: &id002
            type: string
            enum:
            - service
            - level
            - environment
            - release
          maxItems: 2
          uniqueItems: true
        group_limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 10
        histogram:
          type: object
          additionalProperties: false
          required:
          - field
          - interval
          properties:
            field:
              type: string
              enum:
              - timestamp
            interval:
              type: string
              enum:
              - auto
              - 10s
              - 1m
              - 5m
              - 1h
              - 6h
              - 1d
    AggregateMetric:
      oneOf:
      - type: object
        additionalProperties: false
        required:
        - name
        - op
        - value
        - numeric_value_count
        properties:
          name: *id001
          op:
            type: string
            enum:
            - count
          value:
            $ref: '#/components/schemas/NonNegativeDecimalString'
          numeric_value_count:
            type: 'null'
      - type: object
        additionalProperties: false
        required:
        - name
        - op
        - value
        - numeric_value_count
        properties:
          name: *id001
          op:
            type: string
            enum:
            - sum
            - min
            - max
            - avg
          value:
            type:
            - number
            - 'null'
          numeric_value_count:
            oneOf:
            - $ref: '#/components/schemas/NonNegativeDecimalString'
            - type: 'null'
    AggregateBucketKey:
      oneOf:
      - type: object
        additionalProperties: false
        required:
        - type
        - timestamp_us
        properties:
          type:
            type: string
            enum:
            - timestamp
          timestamp_us:
            $ref: '#/components/schemas/Int64String'
      - type: object
        additionalProperties: false
        required:
        - type
        - value
        properties:
          type:
            type: string
            enum:
            - string
          value:
            type: string
    AggregateBucket:
      type: object
      additionalProperties: false
      required:
      - key
      - doc_count
      - metrics
      - children
      properties:
        key:
          $ref: '#/components/schemas/AggregateBucketKey'
        doc_count:
          $ref: '#/components/schemas/NonNegativeDecimalString'
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/AggregateMetric'
        children:
          oneOf:
          - $ref: '#/components/schemas/AggregateBucketSet'
          - type: 'null'
    AggregateBucketSet:
      type: object
      additionalProperties: false
      required:
      - dimension
      - buckets
      - has_more
      properties:
        dimension:
          oneOf:
          - type: object
            additionalProperties: false
            required:
            - histogram
            properties:
              histogram:
                type: object
                additionalProperties: false
                required:
                - interval_ms
                properties:
                  interval_ms:
                    type: integer
                    minimum: 1
          - type: object
            additionalProperties: false
            required:
            - group
            properties:
              group:
                type: object
                additionalProperties: false
                required:
                - field
                properties:
                  field: *id002
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/AggregateBucket'
        has_more:
          type: boolean
    AggregateWarning:
      type: object
      additionalProperties: false
      required:
      - numeric_exclusion_count_unavailable
      properties:
        numeric_exclusion_count_unavailable:
          type: object
          additionalProperties: false
          required:
          - metric
          - field
          properties:
            metric: *id001
            field:
              type: object
              additionalProperties: false
              required:
              - json
              properties:
                json:
                  type: string
    AggregateResponse:
      type: object
      additionalProperties: false
      required:
      - record_count
      - metrics
      - buckets
      - warnings
      - read_token
      - watermark
      - complete
      - took_ms
      - searched_shards
      - hydrated_shards
      properties:
        record_count:
          $ref: '#/components/schemas/NonNegativeDecimalString'
        metrics:
          type: array
          items:
            $ref: '#/components/schemas/AggregateMetric'
        buckets:
          oneOf:
          - $ref: '#/components/schemas/AggregateBucketSet'
          - type: 'null'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/AggregateWarning'
        read_token:
          type: string
        watermark:
          $ref: '#/components/schemas/NonNegativeDecimalString'
        complete:
          type: boolean
          const: true
        took_ms:
          $ref: '#/components/schemas/NonNegativeDecimalString'
        searched_shards:
          $ref: '#/components/schemas/NonNegativeDecimalString'
        hydrated_shards:
          $ref: '#/components/schemas/NonNegativeDecimalString'